// src/types.ts
export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage?: number;
  rating?: number;
  stock?: number;
  brand: string;
  category: string;
  sku: string;
  thumbnail: string;
  images?: Array<string>;
}

export interface Category {
    slug: string;
    name: string;
    url: string;
}